refactor: abstract social auth into actions
